+++ src/main/kotlin/me/bossm0n5t3r/leetcode/deepestLeavesSum/DeepestLeavesSum.kt
@@ -1,0 +1,32 @@
+package me.bossm0n5t3r.leetcode.deepestLeavesSum
+
+import me.bossm0n5t3r.leetcode.utils.TreeNode
+import java.util.LinkedList
+
+class DeepestLeavesSum {
+    class Solution {
+        fun deepestLeavesSum(root: TreeNode?): Int {
+            if (root == null) return 0
+            val queue = LinkedList<TreeNode>()
+            queue.add(root)
+            while (queue.isNotEmpty()) {
+                var curLeavesSum = 0
+                val size = queue.size
+                var isQueued = false
+                repeat(size) {
+                    val cur = queue.poll()
+                    curLeavesSum += cur.`val`
+                    if (!isQueued && (cur.left != null || cur.right != null)) {
+                        isQueued = true
+                    }
+                    if (cur.left != null) queue.add(cur.left!!)
+                    if (cur.right != null) queue.add(cur.right!!)
+                }
+                if (!isQueued) {
+                    return curLeavesSum
+                }
+            }
+            return -1
+        }
+    }
+}
